pyproject.toml
[project]
name = "gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "google-genai>=1.20.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "ruff>=0.11.13",
    "uvicorn>=0.34.3",
]
